Remove file that was 99% duplicated
To be able to share most the file, when the extension loads it fetches
capture-html5-video.js and stores it as a string. After, when a video
screenshot is requested by the user, a copy of the string is modified
to include an implementation for calculating the width and height of
a video according to user settings.